Use pbm for temp images, because they need to be bitonal.
For some reason the ImageMagick tif (and png?) writer writes 8-bit images, even if you say you explicitly tell it you want “-type Bitonal”. That throws the djvu encoder off track. So use formats that are *alawys* bitonal when we have to. (We know the DPI, so it is no problem if the format does not keep track of it.)